attached_assets/requirements_1753667070937.txt

flask
pyyaml
sqlite-utils
